refactor(client): extract LO frequency editor
